refactor: move candidate types to types.ts for consistency
PlayerCandidate and ClubCandidate were defined in helpers.ts while the
other data types (StintFragment, OrphanPlayer, OrphanClub) lived in
types.ts per project convention. Consolidates all types into types.ts
and uses the WinnerSide type alias for pickDefaultWinner's return type.
